+++ docs/source/release_notes.rst
@@ -4,13 +4,13 @@ Release Notes
     * Enhancements
     * Fixes
     * Changes
-        * Remove Decision Tree and CatBoost Estimators from AutoML search :pr:`4205`
     * Documentation Changes
     * Testing Changes
 
 .. warning::
 
     **Breaking Changes**
+        * Removed Decision Tree and CatBoost Estimators from AutoML search :pr:`4205`
 
 
 **v0.77.0 June. 07, 2023**